// src/app/models/covid-data.ts
export class CovidData {
    ID:string;
    Country:string;
    CountryCode:string;
    Province: string;
    City:string;
    CityCode:string;
    Lat: string;
    Lon: string;
    Confirmed: number;
    Deaths: number;
    Recovered: number;
    Active:number;
    Date: Date;
    DailyConfirmed: number;
}